+++ chainer/utils/warnings_.py
